ref: post service delete
